Optimized Db, Autorization, Authentification

+++ src/GraphEditor/GraphEditor/Controllers/GraphController.cs
@@ -1,8 +1,11 @@
-﻿using GraphEditor.Models;
-using System.Text.Json;
+﻿using System.Text.Json;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
 using GraphEditor.Hubs;
+using Microsoft.AspNetCore.Authorization;
+using System.Reflection.Metadata;
+using GraphEditor.Models.Graph;
+using GraphEditor.Models.CRUD;
 
 namespace GraphEditor.Controllers;
 
@@ -10,40 +13,60 @@ namespace GraphEditor.Controllers;
 [Route("/api/v1/[controller]")]
 public class GraphController : Controller
 {
-    private readonly IGraphRepository graphRepository;
+    private readonly IAuthorizationService authorizationService;
+    private readonly IRepository<GraphRecord> graphRepository;
     private readonly IHubContext<GraphHub> hubContext;
 
-    public GraphController(IGraphRepository repo, IHubContext<GraphHub> hub)
+    public GraphController(IAuthorizationService auth,
+                           IRepository<GraphRecord> repo,
+                           IHubContext<GraphHub> hub)
     {
+        authorizationService = auth;
         graphRepository = repo;
         hubContext = hub;
-    }
-
-    [HttpGet("{graphId}")]
-    public async Task<ActionResult> Get(string graphId)
-    {
-        var graph = await graphRepository.Get(graphId);
-        if (graph == null)
-            return NotFound();
-        return Ok();
     }
 
     [HttpPut("{graphId}")]
     public async Task<ActionResult> Create(string graphId)
     {
-        var graph = await graphRepository.Get(graphId);
+        if (User == null)
+            Unauthorized();
+        var graph = await graphRepository.Find(graphId);
         if (graph != null)
             return NoContent();
-        graph = new GraphRecord();
-        var path = ControllerContext.HttpContext.Request.Path;
-        graph.Id = graphId;
-        await graphRepository.Create(graph);
+        graph = new GraphRecord(graphId);        
+        var path = ControllerContext.HttpContext.Request.Path;        
+        await graphRepository.Add(graph);
         return Created(path, JsonSerializer.Serialize(graph));
+    }
+
+    [HttpGet("{graphId}")]
+    public async Task<ActionResult> Get(string graphId)
+    {
+        var graph = await graphRepository.Find(graphId);
+        if (graph == null)
+            return NotFound();
+
+        var authorizationResult = await authorizationService
+            .AuthorizeAsync(User, graph, StringConstants.GraphCRUDPolicy);
+        if (!authorizationResult.Succeeded)
+            return Unauthorized();
+
+        return Ok(graphId);
     }
 
     [HttpDelete("{graphId}")]
     public async Task<ActionResult> Delete(string graphId)
     {
+        var graph = await graphRepository.Find(graphId);
+        if (graph == null)
+            return NotFound();
+
+        var authorizationResult = await authorizationService
+            .AuthorizeAsync(User, graph, StringConstants.GraphCRUDPolicy);
+        if (!authorizationResult.Succeeded)
+            return Unauthorized();
+
         await hubContext.Clients.Group(graphId).SendCoreAsync("OnGraphDelete", Array.Empty<object?>());
         await graphRepository.Delete(graphId);
         return Ok();
